refactor(app): extract event-application orchestrators (phase 6c)
Move log_and_apply_event, apply_event_to_active_buffer,
apply_events_as_bulk_edit, trigger_plugin_hooks_for_event,
handle_scroll_event, handle_set_viewport_event, handle_recenter_event,
invalidate_layouts_for_buffer out of mod.rs into a new
app/event_apply.rs sibling. ~858 lines.

This is the canonical event-application path: every buffer mutation
in the editor flows through one of these methods. Co-located with the
viewport-event handlers (Scroll/SetViewport/Recenter) since
apply_event_to_active_buffer dispatches to them by event type.

mod.rs goes from 6561 -> 5705 lines (-856). 2026 tests still pass.
